[KLT] Dump kernel description.

+++ src/midend/KLT/include/KLT/RTL/kernel.h
@@ -81,7 +81,7 @@ struct klt_kernel_t * klt_build_kernel(int idx);
 
 void klt_execute_kernel(struct klt_kernel_t * kernel);
 
-void klt_dbg_dump_kernel(int idx);
+void klt_dbg_dump_kernel(size_t idx);
 
 void klt_dbg_dump_all_kernels();
